Move social assertion
